+++ src/main/java/Ui.java
@@ -1,0 +1,32 @@
+import java.io.*;
+import java.util.Scanner;
+
+public class Ui {
+    private final BufferedReader in;
+    private final PrintStream out;
+
+    public Ui() {
+        this(System.in, System.out);
+    }
+
+    public Ui(InputStream in, PrintStream out) {
+        this.in = new BufferedReader(
+                new InputStreamReader(in));
+        this.out = out;
+    }
+
+    public void init() {
+        String logo = " ____        _        \n"
+                + "|  _ \\ _   _| | _____ \n"
+                + "| | | | | | | |/ / _ \\\n"
+                + "| |_| | |_| |   <  __/\n"
+                + "|____/ \\__,_|_|\\_\\___|\n";
+        System.out.println("Hello from\n" + logo);
+        System.out.println("Hello! I'm Duke");
+        System.out.println("What can I do for you?");
+    }
+
+    public String getNextCommand() throws IOException {
+        return this.in.readLine();
+    }
+}
